Добавляет модульность, выделяет вспомогательные функции в модуль util.js
